+++ packages/codegen/src/generate-types.ts
@@ -2,7 +2,7 @@ import { anonymizeImports, anonymizeType } from "./anonymize"
 import { CodeDeclarations, Variable } from "./types-builder"
 
 const getTypeDependencies = (input: string) =>
-  new Set([...input.matchAll(/Anonymize<(I\w*)>/gm)].map((x) => x[1]))
+  new Set([...input.matchAll(/Anonymize<(I\w*)>/gm)].map((x) => x[1].slice(1)))
 
 const optimizeVariables = (
   variables: Map<string, Variable>,
@@ -16,39 +16,36 @@ const optimizeVariables = (
   ;[...variables.entries()].forEach(([checksum, { type }]) => {
     const iDependencies = getTypeDependencies(type)
     dependencies.set(checksum, iDependencies)
-    iDependencies.forEach((preChecksum) => {
-      const dependencyChecksum = preChecksum.slice(1)
+    iDependencies.forEach((dependencyChecksum) => {
       const s = dependants.get(dependencyChecksum)
       if (s) {
         s.add(checksum)
         toRemove.delete(dependencyChecksum)
       } else {
         dependants.set(dependencyChecksum, new Set([checksum]))
-        if (!commonTypeImports.has(dependencyChecksum))
+        if (!commonTypeImports.has(variables.get(dependencyChecksum)!.name))
           toRemove.add(dependencyChecksum)
       }
     })
   })
 
-  let roundIdx = 0
   while (true) {
-    console.log({ roundIdx })
     const nonDependants = [...toRemove].filter((item) =>
       [...(dependencies.get(item) ?? [])].every((x) => !toRemove.has(x)),
     )
 
     if (!nonDependants.length) break
 
-    nonDependants.forEach((x) => {
-      const [target] = [...dependants.get(x)!]
-      console.log(target, x)
+    nonDependants.forEach((checksum) => {
+      const variableToInline = variables.get(checksum)!
+      const [target] = [...dependants.get(checksum)!]
       const variable = variables.get(target)!
       variable.type = variable.type.replaceAll(
-        `Anonymize<${x}>`,
-        variables.get(x)!.type,
+        `Anonymize<${variableToInline.name}>`,
+        variableToInline.type,
       )
-      variables.delete(x)
-      toRemove.delete(x)
+      variables.delete(checksum)
+      toRemove.delete(checksum)
     })
   }